# app/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "title": "Log Alarm Management Service",
        "contact": {},
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0"
    },
    "basePath": "/api/v1",
    "paths": {
        "/alert/creat": {
            "post": {
                "security": [
                    {
                        "BasicAuth": []
                    }
                ],
                "description": "创建告警规则",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "alert"
                ],
                "summary": "Creat Alert",
                "parameters": [
                    {
                        "type": "string",
                        "description": "名称",
                        "name": "name",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "索引",
                        "name": "index",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "查询字符串",
                        "name": "query_string",
                        "in": "query",
                        "required": true
                    },
                    {
                        "description": "默认配置",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/main.CreatAlertParamBody"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/alert/delete": {
            "delete": {
                "security": [
                    {
                        "BasicAuth": []
                    }
                ],
                "description": "删除告警规则",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "alert"
                ],
                "summary": "Delete Alert",
                "parameters": [
                    {
                        "type": "string",
                        "description": "名称",
                        "name": "name",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "索引",
                        "name": "index",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "204": {
                        "description": "No Content",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/alert/query": {
            "get": {
                "security": [
                    {
                        "BasicAuth": []
                    }
                ],
                "description": "查询单个索引相关告警",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "alert"
                ],
                "summary": "Query Alerts",
                "parameters": [
                    {
                        "type": "string",
                        "description": "索引",
                        "name": "index",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/monitor/health_check": {
            "get": {
                "description": "健康检查",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "monitor"
                ],
                "summary": "Health Check",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "main.CreatAlertParamBody": {
            "type": "object",
            "required": [
                "delay",
                "description",
                "threshold"
            ],
            "properties": {
                "delay": {
                    "type": "string",
                    "example": "3m"
                },
                "description": {
                    "type": "string",
                    "example": "description"
                },
                "threshold": {
                    "type": "string",
                    "example": "\u003e=10"
                }
            }
        }
    },
    "securityDefinitions": {
        "BasicAuth": {
            "type": "basic"
        }
    }
}